chore: rename user-service to user
