+++ src/redux/user/slice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import type { User } from '@/server/user-service';
+import type { User } from '@/server/user';
 import { join } from '../channel';
 import { login } from './action';
 import { leave } from '@/redux/channel';
